+++ GothicModComposer.UI/ViewModels/GmcVM.cs
@@ -13,6 +13,7 @@ namespace GothicModComposer.UI.ViewModels
     {
         private readonly IGmcExecutor _gmcExecutor;
         private readonly ISpacerService _spacerService;
+        private WindowState _currentWindowState;
 
         public GmcVM(IGmcExecutor gmcExecutor, ISpacerService spacerService, GmcSettingsVM gmcSettingsVM)
         {
@@ -36,6 +37,13 @@ namespace GothicModComposer.UI.ViewModels
             RunSpacer = new RelayCommand(RunSpacerExecute);
             DeleteZenWorld = new RelayCommand(DeleteZenWorldExecute);
             RenameZenWorld = new RelayCommand(RenameZenWorldExecute);
+            ManualZenWorldListRefresh = new RelayCommand(ManualZenWorldListRefreshExecute);
+        }
+
+        public WindowState CurrentWindowState
+        {
+            get => _currentWindowState;
+            set => SetProperty(ref _currentWindowState, value);
         }
 
         public GmcSettingsVM GmcSettings { get; }
@@ -51,6 +59,7 @@ namespace GothicModComposer.UI.ViewModels
         public RelayCommand RunSpacer { get; }
         public RelayCommand DeleteZenWorld { get; }
         public RelayCommand RenameZenWorld { get; }
+        public RelayCommand ManualZenWorldListRefresh { get; }
 
         private void RunUpdateProfileExecute(object obj)
         {
@@ -60,7 +69,11 @@ namespace GothicModComposer.UI.ViewModels
                 return;
             }
 
+            CurrentWindowState = WindowState.Minimized;
+            
             _gmcExecutor.Execute(GmcExecutionProfile.Update);
+            
+            CurrentWindowState = WindowState.Normal;
         }
 
         private void RunComposeProfileExecute(object obj)
@@ -72,9 +85,16 @@ namespace GothicModComposer.UI.ViewModels
             }
 
             var messageBoxResult = MessageBox.Show("Are you sure you want to execute 'Compose' profile?",
-                "Execute Confirmation", MessageBoxButton.YesNo);
+                "Execute Confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            
             if (messageBoxResult == MessageBoxResult.Yes)
+            {
+                CurrentWindowState = WindowState.Minimized;
+            
                 _gmcExecutor.Execute(GmcExecutionProfile.Compose);
+            
+                CurrentWindowState = WindowState.Normal;
+            }
         }
 
         private void RunModProfileExecute(object obj)
@@ -85,15 +105,26 @@ namespace GothicModComposer.UI.ViewModels
                 return;
             }
 
+            CurrentWindowState = WindowState.Minimized;
+            
             _gmcExecutor.Execute(GmcExecutionProfile.RunMod);
+            
+            CurrentWindowState = WindowState.Normal;
         }
 
         private void RunRestoreGothicProfileExecute(object obj)
         {
             var messageBoxResult = MessageBox.Show("Are you sure to you want to execute 'RestoreGothic' profile?",
-                "Execute Confirmation", MessageBoxButton.YesNo);
+                "Execute Confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            
             if (messageBoxResult == MessageBoxResult.Yes)
+            {
+                CurrentWindowState = WindowState.Minimized;
+            
                 _gmcExecutor.Execute(GmcExecutionProfile.RestoreGothic);
+            
+                CurrentWindowState = WindowState.Normal;
+            }
         }
 
         private void RunBuildModFileProfileProfileExecute(object obj)
@@ -104,7 +135,11 @@ namespace GothicModComposer.UI.ViewModels
                 return;
             }
 
+            CurrentWindowState = WindowState.Minimized;
+            
             _gmcExecutor.Execute(GmcExecutionProfile.BuildModFile);
+            
+            CurrentWindowState = WindowState.Normal;
         }
 
         private void RunEnableVDFProfileProfileExecute(object obj)
@@ -182,6 +217,11 @@ namespace GothicModComposer.UI.ViewModels
 
                 File.Move(fullWorldPath, newFileNamePath);
             }
+        }
+
+        private void ManualZenWorldListRefreshExecute(object obj)
+        {
+            GmcSettings.LoadZen3DWorlds(true);
         }
 
         private static void ShowGothicExeNotFoundMessage() =>